Use a long long for element count and fix option handler
Instead of detecting NULL multibulk by comparing with (size_t)-1 just
use a long long.  This also avoids potential issues on 32 bit platforms
since size_t will be 4 bytes, but long long 8.

Note:  Instead of testing for exactly -1, I'm testing if the element
       count is < 0.  Otherwise an invalid length (e.g. -3) would lead
       to undefined behavior.

       It may be prudent to actually do this test but we're going to be
       replacing all parsing logic with Hiredis for PhpRedis 6, which
       handles that correctly.

See: #1847

+++ cluster_library.h
@@ -325,14 +325,14 @@ typedef struct clusterReply {
     size_t integer;                /* Integer reply */
     long long len;                 /* Length of our string */
     char *str;                     /* String reply */
-    size_t elements;               /* Count of array elements */
+    long long elements;            /* Count of array elements */
     struct clusterReply **element; /* Array elements */
 } clusterReply;
 
 /* Direct variant response handler */
 clusterReply *cluster_read_resp(redisCluster *c, int status_strings);
 clusterReply *cluster_read_sock_resp(RedisSock *redis_sock,
-    REDIS_REPLY_TYPE type, char *line_reply, size_t reply_len);
+    REDIS_REPLY_TYPE type, char *line_reply, long long reply_len);
 void cluster_free_reply(clusterReply *reply, int free_data);
 
 /* Cluster distribution helpers for WATCH */
